logic and ui are refactored
