+++ adminUserSite/src/main/java/com/example/adminusersite/servlets/AdminServlet.java
@@ -1,21 +1,19 @@
-package com.example.adminusersite.servlets.crud;
+package com.example.adminusersite.servlets;
+
 
 import com.example.adminusersite.dao.UserDao;
 import com.example.adminusersite.dao.UserDaoImpl;
-import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.SneakyThrows;
 
-@WebServlet("/delete")
-public class DeleteUserServlet extends HttpServlet {
-    @SneakyThrows
+public class AdminServlet extends HttpServlet {
     @Override
+    @SneakyThrows
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
         UserDao userDao = new UserDaoImpl();
-        userDao.deleteUser(Integer.parseInt(req.getParameter("id")));
-
-        req.getRequestDispatcher("sign-in").forward(req, resp);
+        req.setAttribute("dao", userDao);
+        req.getRequestDispatcher("adminPage.jsp").forward(req, resp);
     }
 }
