+++ adminUserSite/src/main/webapp/adminPage.jsp
@@ -1,6 +1,6 @@
 <%@ page import="com.example.adminusersite.dao.UserDao" %>
 <%@ page import="com.example.adminusersite.user.User" %>
-<%@ page contentType="text/html;charset=UTF-8"%>
+<%@ page contentType="text/html;charset=UTF-8" %>
 <html>
 <head>
     <title>Title</title>
@@ -15,12 +15,16 @@
 <div class="wrapper">
     <div class="text-center mt-4 name">
         <div>
+            <div class="text-center mt-4 name">
+                Admin page
+            </div>
             <table>
                 <tr>
                     <td>id</td>
                     <td>name</td>
                     <td>password</td>
                     <td>type</td>
+                    <td>actions</td>
                 </tr>
                 <%
                     for (User u : users) {
@@ -35,6 +39,18 @@
                     </td>
                     <td><%=u.getU_type()%>
                     </td>
+                    <td>
+                        <form method="post" action="delete">
+                            <input type="hidden" name="id" value="<%= u.getId() %>">
+                            <input type="submit" value="delete">
+                        </form>
+                    </td>
+                    <td>
+                        <form method="post" action="update">
+                            <input type="hidden" name="id" value="<%= u.getId() %>">
+                            <input type="submit" value="edit">
+                        </form>
+                    </td>
                 </tr>
                 <%
                     }
@@ -42,8 +58,6 @@
             </table>
             <br>
             <a href="create.jsp">create</a>
-            <a href="delete.jsp">delete</a>
-            <a href="update.jsp">update</a>
         </div>
         <div class="text-center fs-6">
             <a href="index.jsp">start page</a>
