+++ adminUserSite/src/main/webapp/create.jsp
@@ -2,12 +2,24 @@
 <html>
 <head>
     <title>Title</title>
+    <link rel="stylesheet" href="index.css">
 </head>
 <body>
-<form action="register" method="post">
-    login: <label for="login"></label><input type="text" id="login" name="login"><br>
-    password: <label for="password"></label><input type="password" id="password" name="password"><br>
-    <input type="submit" value="register">
-</form>
+<div class="wrapper">
+    <div class="text-center mt-4 name">
+        Registration page
+    </div>
+    <form class="p-3 mt-3" action="create" method="post">
+        <div class="form-field d-flex align-items-center">
+            <span class="far fa-user"></span>
+            <label for="login"></label><input type="text" name="login" id="login" placeholder="Username">
+        </div>
+        <div class="form-field d-flex align-items-center">
+            <span class="fas fa-key"></span>
+            <label for="password"></label><input type="password" name="password" id="password" placeholder="Password">
+        </div>
+        <button class="btn mt-3" type="submit">Registrate</button>
+    </form>
+</div>
 </body>
 </html>
