--- adminUserSite/src/main/webapp/delete.jsp
@@ -1,12 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>Title</title>
-</head>
-<body>
-<form action="delete" method="post">
-    Id: <label for="id"></label><input type="text" id="id" name="id"><br>
-    <input type="submit" value="delete">
-</form>
-</body>
-</html>
