+++ adminUserSite/src/main/webapp/index.css
@@ -14,29 +14,13 @@ body {
 }
 
 .wrapper {
-    max-width: 350px;
-    min-height: 500px;
+    max-width: 500px;
+    min-height: 800px;
     margin: 80px auto;
     padding: 40px 30px 30px 30px;
     background-color: #ecf0f3;
     border-radius: 15px;
     box-shadow: 13px 13px 20px #cbced1, -13px -13px 20px #fff;
-}
-
-.logo {
-    width: 80px;
-    margin: auto;
-}
-
-.logo img {
-    width: 100%;
-    height: 80px;
-    object-fit: cover;
-    border-radius: 50%;
-    box-shadow: 0 0 3px #5f5f5f,
-    0 0 0 5px #ecf0f3,
-    8px 8px 15px #a7aaa7,
-    -8px -8px 15px #fff;
 }
 
 .wrapper .name {
