+++ adminUserSite/src/main/webapp/update.jsp
@@ -2,13 +2,24 @@
 <html>
 <head>
     <title>Title</title>
+    <link rel="stylesheet" href="index.css">
 </head>
+
 <body>
-<form action="update" method="post">
-    new login: <label for="login"></label><input type="text" id="login" name="login"><br>
-    new password: <label for="password"></label><input type="password" id="password" name="password"><br>
-    id: <label for="id"></label><input type="text" id="id" name="id"><br>
-    <input type="submit" value="update">
-</form>
+
+<% var u = request.getAttribute("id");
+%>
+<div class="wrapper">
+    <div class="text-center mt-4 name">
+        <div>
+            <form action="update" method="post">
+                new login: <label for="login"></label><input type="text" id="login" name="login"><br>
+                new password: <label for="password"></label><input type="password" id="password" name="password"><br>
+                <input type="hidden" name="id" value="<%= u %>">
+                <input type="submit" value="update">
+            </form>
+        </div>
+    </div>
+</div>
 </body>
 </html>
